EchoNotes/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
supabase>=2.0.0,<2.13.0
pytest
httpx
